requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib
bcrypt
python-dotenv
mysql-connector-python
python-jose
python-multipart